Rename and move OrientationEnum to Orientation
The enum `OrientationEnum` has been renamed to `Orientation` and relocated to the proper package directory. This improves naming consistency and ensures adherence to the project's package structure.
